FIX: resoling an issue with the last commit that broke the tiff-stack tests
