+++ suitcase/tiff_series/tests/tests.py
@@ -3,6 +3,7 @@ from .. import export
 import numpy
 from numpy.testing import assert_array_equal
 import os
+import pytest
 import tifffile
 
 
@@ -39,13 +40,20 @@ def create_expected(collector, stack_images):
                 expected_dict['baseline'] = numpy.ones((10, 10))
             elif len(event_list) == 1:
                 expected_dict[stream_name] = numpy.ones((10, 10))
-                expected_dict['baseline'] = numpy.ones((2, 10, 10))
+                expected_dict['baseline'] = numpy.ones((2,10, 10))
             else:
                 expected_dict[stream_name] = numpy.ones(
                     (len(event_list), 10, 10))
                 expected_dict['baseline'] = numpy.ones((3, 10, 10))
 
     return expected_dict
+
+
+@pytest.fixture(params=['test'],
+                scope='function')
+def test_path_formatting():
+    ...
+
 
 
 def test_export(tmp_path, example_data):
